# backend/internal/infra/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "个人记账应用后端 API",
        "title": "Personal Bookkeeping API",
        "termsOfService": "https://github.com/your-repo",
        "contact": {
            "name": "Developer",
            "email": "dev@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8000",
    "basePath": "/api/v1",
    "paths": {
...
        "/exchange-rates": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "exchange-rates"
                ],
                "summary": "汇率列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "日期筛选 2006-01-02",
                        "name": "date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "源币种",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "目标币种",
                        "name": "to",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "exchange-rates"
                ],
                "summary": "创建汇率（同日期+币种自动覆盖）",
                "parameters": [
                    {
                        "description": "汇率信息",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.CreateExchangeRateInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handlers.Response"
                        }
                    }
                }
            }
        },
        "/exchange-rates/latest": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "exchange-rates"
                ],
                "summary": "最新汇率（每种币对一条）",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.Response"
                        }
                    }
                }
            }
        },
        "/exchange-rates/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "exchange-rates"
                ],
                "summary": "删除汇率",
                "parameters": [
                    {
                        "type": "string",
                        "description": "汇率 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.Response"
                        }
                    }
                }
            }
        },
        "/ledgers": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ledgers"
                ],
                "summary": "账本列表",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ledgers"
                ],
                "summary": "创建账本",
                "parameters": [
                    {
                        "description": "账本信息",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.CreateLedgerInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handlers.Response"
                        }
                    }
                }
            }
        },
        "/ledgers/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ledgers"
                ],
                "summary": "账本详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "账本 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.Response"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ledgers"
                ],
                "summary": "更新账本",
                "parameters": [
                    {
                        "type": "string",
                        "description": "账本 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新内容",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.UpdateLedgerInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ledgers"
                ],
                "summary": "删除账本",
                "parameters": [
                    {
                        "type": "string",
                        "description": "账本 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.Response"
                        }
                    }
                }
            }
        },
        "/ledgers/{id}/summary": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ledgers"
                ],
                "summary": "账本汇总",
                "parameters": [
                    {
                        "type": "string",
                        "description": "账本 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handlers.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/handlers.ledgerSummary"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handlers.CategorySummary": {
            "type": "object",
            "properties": {
                "category_icon": {
                    "type": "string"
                },
                "category_id": {
                    "type": "string"
                },
                "category_name": {
                    "type": "string"
                },
                "count": {
                    "type": "integer"
                },
                "total": {
                    "type": "number"
                }
            }
        },
...
        "handlers.CreateExchangeRateInput": {
            "type": "object",
            "required": [
                "from_currency",
                "rate",
                "to_currency"
            ],
            "properties": {
                "date": {
                    "type": "string",
                    "example": "2024-01-15"
                },
                "from_currency": {
                    "type": "string",
                    "example": "USD"
                },
                "rate": {
                    "type": "number",
                    "example": 7.24
                },
                "source": {
                    "type": "string",
                    "example": "bank-of-china"
                },
                "to_currency": {
                    "type": "string",
                    "example": "CNY"
                }
            }
        },
        "handlers.CreateLedgerInput": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "base_currency": {
                    "type": "string",
                    "example": "CNY"
                },
                "color": {
                    "type": "string",
                    "example": "#1890ff"
                },
                "description": {
                    "type": "string",
                    "example": "日常收支记录"
                },
                "icon": {
                    "type": "string",
                    "example": "📒"
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "example": "日常账本"
                }
            }
        },
...
        "handlers.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
        "handlers.UpdateLedgerInput": {
            "type": "object",
            "properties": {
                "base_currency": {
                    "type": "string"
                },
                "color": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "icon": {
                    "type": "string"
                },
                "is_archived": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string",
                    "example": "新账本名"
                },
                "sort_order": {
                    "type": "integer"
                }
            }
        },
...
        "handlers.ledgerSummary": {
            "type": "object",
            "properties": {
                "balance": {
                    "type": "number"
                },
                "base_currency": {
                    "type": "string",
                    "example": "CNY"
                },
                "expense_by_category": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handlers.CategorySummary"
                    }
                },
                "total_expense": {
                    "type": "number"
                },
                "total_income": {
                    "type": "number"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT token，格式: Bearer \u003ctoken\u003e",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}